requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
kafka-python==2.0.2
pyspark==3.4.1
matplotlib==3.7.1
seaborn==0.12.2